+++ src/melee/ft/ftcolanim.c
@@ -110,7 +110,7 @@ void ftCo_800BFFAC(Fighter* fp)
 
 bool ftCo_800BFFD0(Fighter* fp, int arg1, int arg2)
 {
-    long tmp;
+    s32 tmp;
     if (arg1 >= 0x7B) {
         arg1 -= (tmp = 0x7B);
         if (lb_800144C8(&fp->x508, Fighter_804D6538, arg1, arg2)) {
